[FIX] move server.js file to main folder
